@@ -8,8 +8,8 @@ import { OrderTableDataSource } from './order-table-datasource';
   styleUrls: ['./order-table.component.css']
 })
 export class OrderTableComponent implements AfterViewInit {
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
   dataSource: OrderTableDataSource;
 
   /** Columns displayed in the table. Columns IDs can be added, removed, or reordered. */
